mobile layout styles improved
