+++ src-frontend/src/features/Layouts/mobile/Layout.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useMemo } from "react";
 import { PanelLeftIcon } from "lucide-react";
+import { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Button } from "@/components/ui/button";
 import {
@@ -7,39 +7,37 @@ import {
   DrawerContainer,
   DrawerTrigger,
 } from "@/components/ui/drawer";
+import { Empty, EmptyContent, EmptyTitle } from "@/components/ui/empty";
 import { SIDEBAR_NAMESPACE, TABS_NAMESPACE } from "@/i18n/resources";
 import { useTabsStore } from "@/stores/tabs-store";
 import { TabPanelDispatcher } from "../../Tabs/components/TabPanels";
 import { NavigationDrawer } from "./NavigationDrawer";
 
-function MobileTaskPanels() {
+function TaskPanels() {
   const { t } = useTranslation(TABS_NAMESPACE);
   const tabs = useTabsStore((state) => state.tabs);
   const activeTabId = useTabsStore((state) => state.activeTabId);
-  const taskTabs = useMemo(
-    () =>
-      tabs
-        .filter((tab) => tab.type === "task")
-        .sort((a, b) => a.createdAt - b.createdAt),
+  const taskTabs = useMemo(() =>
+    tabs
+      .filter((tab) => tab.type === "task")
+      .sort((a, b) => a.createdAt - b.createdAt),
     [tabs],
   );
 
-  const effectiveActiveTabId = taskTabs.some((tab) => tab.id === activeTabId)
-    ? activeTabId
-    : (taskTabs[0]?.id ?? null);
-
-  if (taskTabs.length === 0) {
+  if (activeTabId === null || taskTabs.length === 0) {
     return (
-      <div className="flex h-full items-center justify-center px-4 text-center">
-        <p className="mb-16">{t("tabs.empty.no_tabs_open")}</p>
-      </div>
+      <Empty className="h-full rounded-none">
+        <EmptyContent className="mb-16">
+          <EmptyTitle>{t("tabs.empty.no_tabs_open")}</EmptyTitle>
+        </EmptyContent>
+      </Empty>
     );
   }
 
   return (
     <div className="h-full overflow-hidden bg-layout-tabs-content">
       {taskTabs.map((tab) => {
-        const isActive = tab.id === effectiveActiveTabId;
+        const isActive = tab.id === activeTabId;
         return (
           <div
             role="tabpanel"
@@ -60,6 +58,7 @@ function MobileTaskPanels() {
 
 export function Layout() {
   const { t } = useTranslation(SIDEBAR_NAMESPACE);
+  const [isNavigationOpen, setIsNavigationOpen] = useState(false);
   const tabs = useTabsStore((state) => state.tabs);
   const activeTabId = useTabsStore((state) => state.activeTabId);
   const setActiveTab = useTabsStore((state) => state.setActive);
@@ -81,7 +80,11 @@ export function Layout() {
 
   return (
     <DrawerContainer className="h-dvh min-h-dvh overflow-hidden bg-layout-tabs-content">
-      <Drawer direction="left">
+      <Drawer
+        direction="left"
+        open={isNavigationOpen}
+        onOpenChange={setIsNavigationOpen}
+      >
         <div className="flex h-full flex-col">
           <header className="flex min-h-12 shrink-0 items-center gap-2 border-b bg-layout-tabs-bar px-2 pt-[env(safe-area-inset-top)]">
             <DrawerTrigger asChild>
@@ -99,10 +102,10 @@ export function Layout() {
             </h1>
           </header>
           <main className="min-h-0 flex-1 pb-[env(safe-area-inset-bottom)]">
-            <MobileTaskPanels />
+            <TaskPanels />
           </main>
         </div>
-        <NavigationDrawer />
+        <NavigationDrawer onClose={() => setIsNavigationOpen(false)} />
       </Drawer>
     </DrawerContainer>
   );
