+++ src-frontend/src/features/SideBar/views/WorkspacesView/types.ts
@@ -1,0 +1,1 @@
+export type WorkspaceItemVariant = "current" | "frequent" | "default";
